- OC-8538 - Created a new user service in the Core project
